backend/fastapi/requirements.txt
fastapi>=0.96.0
uvicorn>=0.22.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.6
pydantic>=2.0.0
python-dotenv>=1.0.0
httpx>=0.24.0 
firebase-admin
langchain>=0.0.267
langchain-community>=0.0.10
langchain-core>=0.1.0
